+++ src/Filesystem/DefaultFilesystem.php
@@ -1,14 +1,16 @@
 <?php
 
-namespace Spatie\MediaLibrary;
+namespace Spatie\MediaLibrary\Filesystem;
 
+use Spatie\MediaLibrary\FileManipulator;
 use Spatie\MediaLibrary\Helpers\File;
 use Illuminate\Contracts\Filesystem\Factory;
 use Spatie\MediaLibrary\Events\MediaHasBeenAdded;
+use Spatie\MediaLibrary\Media;
 use Spatie\MediaLibrary\PathGenerator\PathGeneratorFactory;
 use Illuminate\Contracts\Config\Repository as ConfigRepository;
 
-class Filesystem implements FilesystemInterface
+class DefaultFilesystem implements Filesystem
 {
     /** @var \Illuminate\Contracts\Filesystem\Factory */
     protected $filesystem;