+++ src/HasMedia/HasMediaTrait.php
@@ -2,10 +2,10 @@
 
 namespace Spatie\MediaLibrary\HasMedia;
 
+use Spatie\MediaLibrary\Filesystem\Filesystem;
 use Spatie\MediaLibrary\Media;
 use Illuminate\Support\Collection;
 use Spatie\MediaLibrary\MediaRepository;
-use Spatie\MediaLibrary\FilesystemInterface;
 use Spatie\MediaLibrary\Conversion\Conversion;
 use Spatie\MediaLibrary\FileAdder\FileAdderFactory;
 use Spatie\MediaLibrary\HasMedia\Interfaces\HasMedia;
@@ -239,7 +239,7 @@ trait HasMediaTrait
     {
         $this->getMedia($collectionName)
             ->each(function (Media $media) {
-                app(FilesystemInterface::class)->removeFiles($media);
+                app(Filesystem::class)->removeFiles($media);
                 $media->delete();
             });