+++ src/MediaRepository.php
@@ -8,14 +8,10 @@ use Illuminate\Database\Eloquent\Collection as DbCollection;
 
 class MediaRepository
 {
-    /**
-     * @var \Spatie\MediaLibrary\Media
-     */
+    /** @var \Spatie\MediaLibrary\Media */
     protected $model;
 
-    /**
-     * @param \Spatie\MediaLibrary\Media $model
-     */
+    /** @param \Spatie\MediaLibrary\Media $model */
     public function __construct(Media $model)
     {
         $this->model = $model;
@@ -64,25 +60,16 @@ class MediaRepository
         return $this->model->all();
     }
 
-    /*
-     * Get all media for the given type.
-     */
     public function getByModelType(string $modelType) : DbCollection
     {
         return $this->model->where('model_type', $modelType)->get();
     }
 
-    /*
-     * Get media by ids.
-     */
     public function getByIds(array $ids) : DbCollection
     {
         return $this->model->whereIn('id', $ids)->get();
     }
 
-    /*
-     * Get all media for the given type and collection name.
-     */
     public function getByModelTypeAndCollectionName(string $modelType, string $collectionName) : DbCollection
     {
         return $this->model
@@ -91,9 +78,6 @@ class MediaRepository
             ->get();
     }
 
-    /*
-     * Get all media for the given type and collection name.
-     */
     public function getByCollectionName(string $collectionName) : DbCollection
     {
         return $this->model